requirements.txt
Flask>=3.0.0
selenium>=4.20.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
google-generativeai>=0.7.2
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
